requirements.txt
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
# torch==2.5.1
# torchvision==0.20.1
torch==2.4.1+cu124
# torchaudio==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
tzdata==2024.2
